backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend service for practice-robot"
authors = [
    {name = "tetsu"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "grpcio (>=1.71.0,<2.0.0)",
    "grpcio-tools (>=1.71.0,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"

